requirement.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.6.4
drfdocs==0.0.11
httpie==1.0.2
idna==2.8
psycopg2==2.7.7
psycopg2-binary==2.7.7
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
